More Updates

+++ server/controllers/productController.js
@@ -1,5 +1,6 @@
 import Product from "../models/productModel.js";
 import cloudinary from "../config/cloudinary.js";
+import mongoose from "mongoose";
 
 
 export const getProducts = async (req, res) => {
@@ -13,9 +14,10 @@ export const getProducts = async (req, res) => {
 
 export const getProduct = async (req, res) => {
     const {_id} = req.params
+    console.log(_id)
     try {
         const product = await Product.findById(_id)
-        res.status(200).json(product)
+        res.status(200).json({product})
     }catch(error) {
         res.status(400).json({error: error.message})
     }
@@ -66,7 +68,14 @@ export const updateProduct = async (req, res) => {
     const image = req.file.path;
 
     try {
+        if(!mongoose.Types.ObjectId.isValid(_id)){
+            return res.status(400).json({error: "Invalid product id"})
+        }
+
         const product = await Product.findById(_id);
+        if (product.owner.toString() !== req.user._id.toString()) {
+            return res.status(404).json({error: "You are not the owner of this product"})
+        }
 
         if (req.file) {
             await cloudinary.uploader.destroy(product.imageID)
@@ -98,12 +107,16 @@ export const updateProduct = async (req, res) => {
     }
 }
 
-export const deleteProduct = async (res, req) => {
+export const deleteProduct = async (req, res) => {
     const {_id} = req.params
     if(!mongoose.Types.ObjectId.isValid(_id)){
         return res.status(400).json({error: "Invalid product id"})
     }
     try{
+        const prodOwner = await Product.findById(_id)
+        if (prodOwner.owner.toString() !== req.user._id.toString()) {
+            return res.status(404).json({error: "You are not the owner of this product"})
+        }
         const product = await Product.findByIdAndDelete(_id)
         if(!product){
             return res.status(404).json({error: "product not found"})
